tests/robustness/requirements.txt
# Requirements for prompt robustness testing framework
# Install with: uv add --dev <package>

# Core testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Semantic similarity
sentence-transformers>=2.2.2
transformers>=4.30.0

# Image comparison
Pillow>=10.0.0
ImageHash>=4.3.1
scikit-image>=0.21.0

# Data analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# NLP (optional, for enhanced entity extraction)
# spacy>=3.6.0

# Visualization comparison
# opencv-python>=4.8.0  # Alternative to scikit-image

# Configuration
PyYAML>=6.0.1

# Logging & utilities
coloredlogs>=15.0.1
tqdm>=4.66.0
